LegalDocuSum/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
google-genai==1.0.0
Werkzeug==3.0.1
docx
Flask
flask-cors
google-genai
PyPDF2
python-docx
python-dotenv
sift-stack-py
